move url and authorization exceptions to separate directories
