+++ src/main/java/edu/kpi/testcourse/exception/url/InvalidUrlException.java
@@ -1,4 +1,4 @@
-package edu.kpi.testcourse.exception;
+package edu.kpi.testcourse.exception.url;
 
 /**
  * Throw this exception if user provides a wrong URL to shorten.
